Italicize System Messages
Add a field to the append method in index.html to specify the style
class.
Consider make inline nicknames (in messages) bold, but that would
make things messy or change the structure. Adding a separate field
for nickname (instead of concatenating it server-side) would still
leave some security in that the server is telling the client which
nickname sent that message. Still, the append method would become
less generalized to support separately-styled spans.
Calling jQuery's .addClass() with an undefined parameter seems to
be fine (in the append method).