Pending Messages
Homework 3: Don’t send the same message to the user that sent it
himself. Instead, append the message directly as soon as he presses
enter.
To achieve this functionality as described, one would use the
socket.broadcast.emit() method on the server. However, I took this in a
different direction. When the user sends a message, it is added to his
or her message list immediately, but in a pending state. This appears
as italics so the user knows it has not yet been sent to everyone else.
When the server rebroadcasts it, the client has confirmation and
removes the pending class. All other clients simply add it to their
list. Normally messages will appear italicized for a fraction of a
second before they straighten. I think this is a good thing, as it
makes it clear to the user that their message did, in fact, go through,
that they aren't talking to themselves. Opening a local client and
external connection open side-by-side, I was able to test this by
turning of my Wi-Fi adapter. Messages remain italicized until Wi-Fi is
re-enabled. Furthermore, this handles multiple unsent messages with the
same content, and users with the same username sending a message with
the same content. This is because the receive method (on 'chat
message') only removes the pending status from the first message that
matches, not all. This avoids the appalling idea of global message IDs
mentioned in the previous commit, that would cause more problems than
these edge cases are worth. That being said, I am very happy with how
simple providing this functionally turned out being. No changes were
made in index.js.